feat: add command for simplified deployment (#4)

+++ src/commands/build/index.ts
@@ -17,20 +17,29 @@ export function builder(yargs: Argv) {
             type: 'string',
             default: '.artifacts',
         })
+        .option('target', {
+            type: 'array',
+            default: ['*'],
+            string: true,
+        })
 }
 
 export async function handler(argv: ReturnType<typeof builder>['argv']): Promise<void> {
-    const { buildDir: _buildDir, artifactDir: _artifactDir } = await argv
+    const { buildDir: _buildDir, artifactDir: _artifactDir, target } = await argv
     const buildDir = path.join(rootDirectory, _buildDir)
     const artifactDir = path.join(rootDirectory, _artifactDir)
 
-    await fs.promises.rm(artifactDir, { recursive: true }).catch(() => void {})
-    await fs.promises.mkdir(artifactDir)
+    const targets = target.flatMap((t) => t.split(','))
+
+    if (targets.includes('*')) {
+        await fs.promises.rm(artifactDir, { recursive: true }).catch(() => void {})
+    }
+    await fs.promises.mkdir(artifactDir).catch(() => void {})
 
     const stacks = await globby(['src/**/functions'], { cwd: rootDirectory, onlyDirectories: true })
-    const handlers = (
-        await Promise.all(stacks.flatMap(async (fnDir) => listLambdaHandlers(path.join(rootDirectory, fnDir))))
-    ).flat()
+    const handlers = (await Promise.all(stacks.flatMap(async (fnDir) => listLambdaHandlers(path.join(rootDirectory, fnDir)))))
+        .flat()
+        .filter((handler) => targets.includes('*') || targets.some((t) => handler.includes(t)))
 
     const outbase = path.join(rootDirectory, 'src')
